pyproject.toml
[project]
name = "helm-values-manager"
version = "1.0.0"
description = "A schema-driven tool for managing Helm values with environment-specific configurations and secret management"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.16.0",
    "rich>=14.0.0",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
helm-values-manager = "helm_values_manager.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "tox>=4.0.0",
    "tox-uv>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=helm_values_manager --cov-report=html --cov-report=term"
